Be more tolerant of FlowModel class changes for local dev
When FlowModel classes are changed during dev persisted instances may
fail to deserialize. Furthermore, since the entire DB file was
deserialized at once this makes the entire file unusable and requires
deleting it by hand.

This change makes deserialization more tolerant of failures. If an entry
cannot be deserialized it will be removed but other instances will be
unaffected.