PTCD-305 Remove redirect for pending migrations
The "continue to course directory" link on the migration report was
redirected back to itself via HomeController/index.

Cathie> The link states the user should go back to the Course Directory
homepage.  Once the user has returned to this page, they can return to
the Migration report via the link in the blue box to resolve the issues
of those courses pending.  The user should be able to add more courses
as required - it could be that they don’t have the required information
to resolve those courses that require fixing, but this should not stop
them from adding/amending other course information. As long as they can
return to the Migration report having logged in/out to resolve these
issues, then it will be fine.

There is no test coverage at the moment, and we have not yet agreed to
retrofit it here so leaving bare for now.

Co-authored-by: MrKevJoy <MrKevinJoy@outlook.com>