PTCD-508 - Set delivery_mode to employer in bulk upload
Set delivery_mode to employer if delivery_method is employer. This is the change in behaviour requested in the jira ticket:

> "Current apprenticeship bulk upload implementation will always have delivery modes empty for a location that is employer based. This is different behaviour to the non-bulk upload journey, which sets the delivery modes to Employer. Change required to make bulk upload on apprenticeships to include 'employer' as the delivery mode."

This follows the existing pattern of mutating data in the `Mandatory_Checks_*` methods.

This is the minimum patch needed to achieve the requested behaviour change in order to separate it from refactorings to improve the code.