Use IProviderContextProvider everywhere
Previously lots of components fished out the provider context from the
feature on HttpContext rather than going through this type. Setting the
context was done in a similar way. Move all gets & sets to use this
abstraction instead.

This also moves the legacy session-based fallback into the provider and
out of the Resource Filter. It's important that this fallback occurs
'late' otherwise having multiple pages open for different providers will
not work as expected.