Add explicit clock to lARS import test
This value was an important "mystery guest" to the tests.

The framework records are filtered by the dates, so the test numbers
would be affected by changes to the base clock's default time.

This time is a copy of the base clock's default time so that this patch
does not alter the test behaviour.