Amend how the 'provider context' ID is appended in Razor views
Previously we had a magic tag helper that would run *after* the href was
generated that appended the query param. This technique won't work on
some of the GDS tag helpers, however (since some are 'renderless').

The new approach uses the 'asp-all-route-data' attribute common to all
link-generating tag helpers (including the GDS ones) and a helper class
to get the values.