Update UkrlpSyncHelper and the UKRLP tests for the v6 XML
Update data mappings for UKRLP v6 sync. This fixes the compilation of
production code. - This change in this patch makes the minimal possible
change to the shape of the output data in order to minimize impact on
downstream systems.

The updated XML in the tests is a trimmed example from the v6 live
service with contact details still anonymised.