Don't test UKRLP time deserialization
Test failed on CI due to some sort of timezone difference on the CI servers.
The rabbit hole with timezones goes deep and it's not worth the time when the fields aren't used anywhere.

Research so far shows that the value from the xml (represented as ISO 8601 local time with +1 offset) is being converted to system local time on read, which will vary depending on the system.
For more fun and games read https://codeblog.jonskeet.uk/2012/05/02/more-fun-with-datetime/
If you have to fix this at some point, good luck.