Update action name in legacy redirect to publish page
Surprisingly it worked without this patch, it appears to fall back to
just generating a url from the parts if it can't find a controller
action, and because the legacy route has been maintained it still
magically worked.

This patch draws attention to the source location that is redirecting to
the new action and removes the misleading specifing of a
controller/action that doesn't exist.

Thanks JL for spotting this at review.