Db command signature consistency
Some of the older DB commands that update data have the signature
IDbCommand<Success> and they either explode if the specified thing
doesn't exist or silently do nothing.

For new commands we've settled on IDbCommand<OneOf<NotFound, Success>>
so we can consistently return the 'not found'-ness.

This commit brings the older commands into line with the new ones.