Simplify invalid state-type errors in v2
Previously each type of error defined its own struct to allow for
passing additional data for each error type. In practice this was never
used. Indeed, only the exception type really matters (as it's mapped to
a 400 by an exception filter later on).

This simplifies things by removing the structs and defining an enum
which can optionally be passed to InvalidStateException.