Cope gracefully with missing venue
We have some bad data that means venueIds are referenced that don't
exist. The venue endpoint returns an empty body, which then becomes
null, which then throws a null-ref exception which isn't very helpful.

By just not setting the value if it's null the page now loads and the
user can correct the corrupted data.

Also corrected use of async.