Always assign legacy provider context where possible
The AssignLegacyProviderContextAttribute exists to ensure that
transitioning from a v2 page (which doesn't require the session-based
provider context) to a legacy page (which does) will work correctly and
the 'ambient' provider is correct.

This attribute has been sprinkled over selected actions where there have
been visible bugs. However, given that the layout produces links to
legacy pages we should *always* been assigning this where we can. This
commit makes that change.

The AssignLegacyProviderContextAttribute is kept around for the day when
the layout no longer links to any legacy pages. At such a time this
change can be reverted.