Replace FindACourseIndex view with a table
Add a stored proc parameterised by course run IDs that refreshes
FindACourseIndex table for those course runs. Run this as part of
Cosmos->SQL sync.

The Azure Search indexer requires that views it targets never have rows
removed, otherwise it cannot track the removal. Instead a 'soft delete'
column is added then search queries have to append a filter to remove
deleted rows from the results.

Achieving this is a challenge for this view; each course run has either
one entry with an ID matching the course run ID, when it has no
sub-regions OR it has a row per subregion, where each row's ID has the
sub region ID appended. When a course run flips from being region-based
or vice-versa then we end up 'losing' the old row and the search indexer
cannot see the removal.

In addition, this view is complex enough that it can't be indexed so
reads from it are not performant.

Moving to a table allows us to add an index so it can be queried
efficiently. The query to populate it can be expressed naturally and the
MERGE in the SPROC takes care of the soft deleting, keeping the
indexer's results correct.