PTCD-823 - Add a new SQL-based provider search index
This is index not being kept compatible with the existing search index
so when we switch calls over to this the models will need to be updated.
I think the switch to a replacement index is a convenient moment to
eliminate the arbitrary variation in naming, unused fields, and
confusing mappings.

Note that the "status" field names are confusing, in this version I've
opted to keep the names aligned with those in the sql table as I think
that will be cleaner in the end.

This replacement index has all the fields needed to display the search
results for admin users at `/provider-search?SearchQuery=xxx`

Information on the relevant CosmosDB to SQL mappings can be found in
Dfc.CourseDirectory.Core\SqlDataSync.cs lines 193-230

The intention is to deploy the two indexes side-by-side, move the
calling code over, and then remove the old index.

Using `assumeOrderByHighWaterMarkColumn` for SQL indexing as per
Dfc.CourseDirectory.FindACourseApi/Resources/AzureSearch/SearchIndexer.json:12
which already works. `convertHighWaterMarkToRowVersion` is for CosmosDB.
See https://docs.microsoft.com/en-us/azure/search/search-howto-connecting-azure-sql-database-to-azure-search-using-indexers#converthighwatermarktorowversion
vs https://docs.microsoft.com/en-us/azure/search/search-howto-index-cosmosdb#incremental-progress-and-custom-queries