Add more unused provider search fields to fix scoring
It seems that just adding the name fields wasn't enough. The guess is
that all the other unused fields are also skewing the numbers because
the old index didn't have a specific list of fields that was searchable.

The field definitions in the search index are exactly as in the original
index definition.