Sync Venue adds & updates to SQL immediately after writing to Cosmos
There are a few places where we read Venues from SQL instead of Cosmos
and if we rely on the background sync process they may not be there by
the time the SQL read happens. This change syncs Venue adds and updates
immediately to avoid this problem in the same way we do for Venue
deletes.