PTCD-964 Decouple fileprocessor
As part of moving away from Cosmos & micro-services we need to remove
usage of the fileprocessor. The sole remaining use of this is for large
FE Bulk Upload files. This commit prevents fileprocessor from processing
these large files (by appending a suffix to files uploaded to Blob
Storage) and instead does the processing in-proc, using our background
worker bits.

The code is the bare minimum to make this work only, given that all this
is going away soon.