Support partial validation of Course Upload Rows
For efficiency we don't want to validate every single row in the upload
more often than we have to. This change amends the validation method to
support changing a subset of rows.

To support this, the grouping of rows into courses is moved out of the
validation method and into ProcessCourseFile (so they will never be
re-grouped after the initial processing).

The LastValidated property is removed from CourseUpload since we no
longer always validate the entire file.