Add support for checking a single entry in ModelState is valid
The RequireValidModelStateAttribute is used for returning a generic
error page if request parameters are not valid (e.g. an enum bound from
a query parameter). However, for POSTs we don't want to necessarily
block all requests if ModelState is not valid; we want them to flow to
our validators. This commit adds a constructor overload to
RequireValidModelStateAttribute where a single key can be specified. In
this case only that ModelStateEntry for the specified key is checked for
validity.