Rationalize HTML encoded course data (#2152)
The legacy UI stores string fields HTML-encoded in Cosmos. As part of the move away from Cosmos we want to move away from encoding data at rest and encoding it 'on the edge' as appropriate. 

This PR adds an additional field to the `Courses` and `CourseRuns` SQL tables indicating whether that row has its strings encoded or not.

All existing data will have the flag set to `1` since it's currently all synced from Cosmos. As we move the UI to using SQL directly we will store things unencoded and the flag will be set to `0`. The new Course Data Management features store fields unencoded also.

The mapping to `Course` and `CourseRun` models is updated here to handle both cases; where the data is encoded and where it's not.