Move FAC index updates to the background for Data Management
Updating the index synchronously on publishing is causing timeouts with
large-ish course counts. This change moves the FAC index updates into
its own transaction that runs separately to publishing. It also updates
rows in relatively small batches to keep the transaction size down.

Since we're now introducing the possibility that publishing succeeds but
the FAC index fails, a Function has been added that runs on a timer to
'mop-up' any index updates that failed in the background task.